clean up git merge
